Renames CaPOYO -> Calcium POYOPlus.
