+++ examples/poyo_plus/configs/finetune.yaml
@@ -1,7 +1,7 @@
 defaults:
   - defaults.yaml
-  - model: capoyo.yaml
-  - dataset: capoyo_single_session.yaml
+  - model: calcium_poyo_plus.yaml
+  - dataset: calcium_poyo_plus_single_session.yaml
   - _self_
 
 batch_size: 64
@@ -14,7 +14,7 @@ train_transforms:
     peak: 4
 
 wandb:
-  run_name: capoyo_finetune
+  run_name: calcium_poyo_plus_finetune
 
 epochs: 100
 ckpt_path: null  # Must specify path to model to finetune
