+++ examples/poyo_plus/configs/model/calcium_poyo_plus_single_session.yaml
@@ -1,4 +1,4 @@
-_target_: torch_brain.models.CaPOYO
+_target_: torch_brain.models.CalciumPOYOPlus
 sequence_length: 1.0
 latent_step: 0.125
 num_latents_per_step: 16
