+++ examples/poyo_plus/configs/train_calcium_poyo_plus.yaml
@@ -1,7 +1,7 @@
 defaults:
   - defaults.yaml
-  - model: capoyo.yaml
-  - dataset: capoyo.yaml
+  - model: calcium_poyo_plus.yaml
+  - dataset: calcium_poyo_plus.yaml
   - _self_  # ensures current config overrides defaults
 
 epochs: 2000
@@ -16,5 +16,5 @@ train_transforms:
     tail_right: 100
 
 wandb:
-  run_name: capoyo_full
+  run_name: calcium_poyo_plus_full
