+++ torch_brain/models/__init__.py
@@ -1,8 +1,8 @@
 from .poyo import POYO, poyo_mp
 from .poyo_plus import POYOPlus
-from .capoyo import CaPOYO
+from .calcium_poyo_plus import CalciumPOYOPlus
 
-__all__ = ["POYO", "POYOPlus", "CaPOYO", "poyo_mp"]
+__all__ = ["POYO", "POYOPlus", "CalciumPOYOPlus", "poyo_mp"]
 
 # see docs/source/api_reference.py
 __api_ref__ = {
